# Conflicts:
#	3rd/ant

+++ packages/gamerender/saveload.lua
@@ -13,16 +13,20 @@ local startup_entities = import_package("vaststars.prototype")("item.startup").e
 local objects = require "objects"
 local ifluid = require "gameplay.interface.fluid"
 local iscience = require "gameplay.interface.science"
+local iui = ecs.import.interface "vaststars.gamerender|iui"
+local GAMEPLAY_VERSION <const> = require "version"
 
 local irq = ecs.import.interface "ant.render|irenderqueue"
 local iom = ecs.import.interface "ant.objcontroller|iobj_motion"
 local ic = ecs.import.interface "ant.camera|icamera"
 local math3d = require "math3d"
 local iobject = ecs.require "object"
+local terrain = ecs.require "terrain"
+local camera = ecs.require "engine.camera"
 
 local MAX_ARCHIVING_COUNT <const> = 9
 
-local archival_relative_dir_list = {}
+local archival_list = {}
 
 local function restore_world()
     -- clean
@@ -31,16 +35,7 @@ local function restore_world()
     end
     objects:clear()
 
-    local duplicate = {}
-    local network_id = 0
-    local function get_network_id(id)
-        if not duplicate[id] then
-            network_id = network_id + 1
-            duplicate[id] = network_id
-        end
-        return duplicate[id]
-    end
-
+    -- restore world
     local function restore_object(gameplay_eid, prototype_name, dir, x, y, fluid_name, fluidflow_network_id)
         local typeobject = iprototype.queryByName("entity", prototype_name)
 
@@ -59,17 +54,19 @@ local function restore_world()
     end
 
     -- restore
+    local all_object = {}
+    local empty_fluidbox = {}
     for v in gameplay_core.select("id:in entity:in fluidbox?in fluidboxes?in") do
         local e = v.entity
         local typeobject = iprototype.queryById(e.prototype)
         local fluid_name = ""
-        local fluidflow_network_id = 0
         if v.fluidbox then
+            fluid_name = ""
             if v.fluidbox.fluid ~= 0 then
                 local typeobject_fluid = assert(iprototype.queryById(v.fluidbox.fluid))
                 fluid_name = typeobject_fluid.name
             else
-                fluidflow_network_id = get_network_id(v.fluidbox.id)
+                empty_fluidbox[iprototype.packcoord(e.x, e.y)] = v.id
             end
         end
         if v.fluidboxes then
@@ -87,8 +84,49 @@ local function restore_world()
                 end
             end
         end
-        restore_object(v.id, typeobject.name, iprototype.dir_tostring(e.direction), e.x, e.y, fluid_name, fluidflow_network_id)
+        all_object[v.id] = {
+            name = typeobject.name, 
+            dir = iprototype.dir_tostring(e.direction),
+            x = e.x,
+            y = e.y,
+            fluid_name = fluid_name,
+            fluidflow_network_id = 0,
+        }
     end
+
+    local function dfs(all_object, empty_fluidbox, object, input_dir)
+        local typeobject = iprototype.queryByName("entity", object.name)
+        for _, v in ipairs(ifluid:get_fluidbox(typeobject.name, object.x, object.y, object.dir)) do
+            if v.dir ~= input_dir then
+                local succ, x, y = terrain:move_coord(v.x, v.y, v.dir, 1)
+                assert(succ)
+                local neighbor_id = empty_fluidbox[iprototype.packcoord(x, y)]
+                if neighbor_id then
+                    local neighbor = assert(all_object[neighbor_id])
+                    if neighbor then
+                        assert(neighbor.fluidflow_network_id == 0)
+                        neighbor.fluidflow_network_id = object.fluidflow_network_id
+                        dfs(all_object, empty_fluidbox, neighbor, iprototype.opposite_dir(v.dir))
+                    end
+                end
+            end
+        end
+    end
+
+    local network_id = 0
+    for _, id in pairs(empty_fluidbox) do
+        local object = all_object[id]
+        if object.fluidflow_network_id == 0 then
+            network_id = network_id + 1
+            object.fluidflow_network_id = network_id
+            dfs(all_object, empty_fluidbox, object)
+        end
+    end
+
+    for id, v in pairs(all_object) do
+        restore_object(id, v.name, v.dir, v.x, v.y, v.fluid_name, v.fluidflow_network_id)
+    end
+
     iobject.flush()
     iscience.update_tech_list(gameplay_core.get_world())
 end
@@ -120,10 +158,15 @@ local function restore_camera_setting(camera_setting)
     ic.set_frustum(ce, camera_setting.frustum)
 end
 
-local M = {}
+local M = {running = false}
 function M:backup()
-    while #archival_relative_dir_list + 1 > MAX_ARCHIVING_COUNT do
-        local archival_relative_dir = table.remove(archival_relative_dir_list, 1)
+    if not self.running then
+        log.error("not running")
+        return
+    end
+
+    while #archival_list + 1 > MAX_ARCHIVING_COUNT do
+        local archival_relative_dir = table.remove(archival_list, 1)
         local archival_dir = archival_base_dir .. ("/%s"):format(archival_relative_dir)
         print("remove", archival_dir)
         fs.remove_all(archival_dir)
@@ -133,10 +176,10 @@ function M:backup()
     local dn = ("%04d-%02d-%02d-%02d-%02d-%02d"):format(t.year, t.month, t.day, t.hour, t.min, t.sec)
     local archival_dir = archival_base_dir .. ("/%s"):format(dn)
 
-    archival_relative_dir_list[#archival_relative_dir_list + 1] = dn
+    archival_list[#archival_list + 1] = {dir = dn, version = require("version")}
     gameplay_core.backup(archival_dir)
 
-    writeall(archiving_list_path, json.encode(archival_relative_dir_list))
+    writeall(archiving_list_path, json.encode(archival_list))
     writeall(camera_setting_path, json.encode(get_camera_setting()))
     print("save success", archival_dir)
 end
@@ -150,51 +193,86 @@ function M:restore(index)
     --
     if not fs.exists(fs.path(archiving_list_path)) then
         self:restart()
-        return
+        self.running = true
+        return true
     end
-    archival_relative_dir_list = json.decode(readall(archiving_list_path))
-    if #archival_relative_dir_list <= 0 then
+    archival_list = json.decode(readall(archiving_list_path))
+    if #archival_list <= 0 then
         self:restart()
-        return
+        self.running = true
+        return true
     end
 
-    index = index or #archival_relative_dir_list
-    if index > #archival_relative_dir_list then
+    index = index or #archival_list
+    if index > #archival_list then
         log.error(("Failed to restore `%s`"):format(index))
-        return
+        iui.open("option_pop.rml")
+        return false
     end
 
     local archival_dir
     while index > 0 do
-        local archival_relative_dir = archival_relative_dir_list[index]
+        local archival_relative_dir = archival_list[index].dir
         archival_dir = archival_base_dir .. ("/%s"):format(archival_relative_dir)
 
         if not fs.exists(fs.path(archival_dir)) then
             log.warn(("`%s` not exists"):format(archival_relative_dir))
-            archival_relative_dir_list[index] = nil
+            archival_list[index] = nil
             index = index - 1
+            goto continue
+        end
+
+        if archival_list[index].version ~= GAMEPLAY_VERSION then
+            log.error(("Failed `%s` version `%s` current `%s`"):format(archival_relative_dir, archival_list[index].version, GAMEPLAY_VERSION))
+            -- archival_list[index] = nil
+            -- index = index - 1
+            iui.open("option_pop.rml")
+            return false
         else
             break
         end
+
+        ::continue::
     end
 
     if index == 0 then
         log.error("Failed to restore")
-        return
+        iui.open("option_pop.rml")
+        return false
     end
 
     gameplay_core.restore(archival_dir)
-    gameplay_core.build()
-    restore_world()
+    self.running = true
+    iscience.update_tech_list(gameplay_core.get_world())
+    iui.open("construct.rml")
 
     if camera_setting then
         restore_camera_setting(camera_setting)
     end
+
+    local coord = terrain:adjust_position(camera.get_central_position(), terrain.ground_width, terrain.ground_height)
+    if coord then
+        terrain:enable_terrain(coord[1], coord[2])
+    end
+
+    gameplay_core.build()
+    restore_world()
+
     print("restore success", archival_dir)
+    return true
 end
 
 function M:restart()
     gameplay_core.restart()
+
+    self.running = true
+    iscience.update_tech_list(gameplay_core.get_world())
+    iui.open("construct.rml")
+
+    local coord = terrain:adjust_position(camera.get_central_position(), terrain.ground_width, terrain.ground_height)
+    if coord then
+        terrain:enable_terrain(coord[1], coord[2])
+    end
 
     for _, e in ipairs(startup_entities) do
         gameplay_core.create_entity(e)
@@ -203,8 +281,8 @@ function M:restart()
     restore_world()
 end
 
-function M:get_archival_relative_dir_list()
-    return archival_relative_dir_list
+function M:get_archival_list()
+    return archival_list
 end
 
 return M
